% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{ADaMData}
\alias{ADaMData}
\title{ADaMData Class}
\usage{
ADaMData(
  data = structure(list(), names = character(0), row.names = integer(0), class =
    "data.frame"),
  domain = "",
  population = "FAS",
  subject_var = "USUBJID",
  trt_var = "TRT01P",
  metadata = list()
)
}
\description{
Base class for ADaM datasets with population filters.
}
