% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{add_table}
\alias{add_table}
\alias{add_plot}
\alias{add_section}
\alias{add_content}
\title{Add a table to a StudyResult}
\usage{
add_table(obj, table, name = NULL)

add_plot(obj, plot, name = NULL)

add_section(obj, section, name = NULL)

add_content(obj, content, name = NULL)
}
\arguments{
\item{obj}{A ReportSection object}

\item{table}{A ClinicalTable object}

\item{name}{Character string for content name/identifier}

\item{plot}{A ClinicalPlot object}

\item{section}{A ReportSection object}

\item{content}{A ClinicalContent object}
}
\value{
The modified StudyResult object

The modified StudyResult object

The modified ClinicalReport object

The modified ReportSection object
}
\description{
Add a table to a StudyResult

Add a plot to a StudyResult

Add a section to a ClinicalReport

Add content to a ReportSection
}
