% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_layer}
\alias{add_layer}
\title{Add a layer to a LayeredTable}
\usage{
add_layer(table, layer)
}
\arguments{
\item{table}{A LayeredTable object}

\item{layer}{An AnalysisLayer object}
}
\value{
Updated LayeredTable
}
\description{
Add a layer to a LayeredTable
}
