% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{AnalysisResults}
\alias{AnalysisResults}
\title{AnalysisResults Class}
\usage{
AnalysisResults(
  stats = structure(list(), names = character(0), row.names = integer(0), class =
    "data.frame"),
  type = "",
  groupings = list(),
  metadata = list()
)
}
\description{
Container for pre-calculated statistics.
}
