% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{analyze}
\alias{analyze}
\title{Analyze ADaM datasets}
\usage{
analyze(x, ...)
}
\arguments{
\item{x}{An ADaMData object or data frame}

\item{...}{Additional arguments passed to methods}
}
\value{
An AnalysisResults object
}
\description{
Generic function for performing clinical analysis on ADaM data.
}
