% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_core.R
\name{analyze_ADaMData}
\alias{analyze_ADaMData}
\title{Analyze ADaMData}
\usage{
analyze_ADaMData(x, ...)
}
\arguments{
\item{x}{An ADaMData object}

\item{...}{Additional arguments}
}
\value{
An AnalysisResults object
}
\description{
S7 method for analyzing ADaMData objects.
}
