% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_core.R
\name{analyze_soc_pt}
\alias{analyze_soc_pt}
\title{Analyze Adverse Events by SOC and PT}
\usage{
analyze_soc_pt(data, soc_var = "AEBODSYS", pt_var = "AEDECOD")
}
\arguments{
\item{data}{ADAE dataset (or ADaMData containing it)}

\item{soc_var}{Character string for SOC variable. Default "AEBODSYS"}

\item{pt_var}{Character string for PT variable. Default "AEDECOD"}
}
\value{
AnalysisResults object
}
\description{
Analyze Adverse Events by SOC and PT
}
