% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_logic.R
\name{analyze_study_OneArmStudy}
\alias{analyze_study_OneArmStudy}
\title{Analyze OneArmStudy}
\usage{
analyze_study_OneArmStudy(x, ...)
}
\arguments{
\item{x}{A OneArmStudy object}

\item{...}{Additional arguments}
}
\value{
The modified OneArmStudy object with results
}
\description{
S7 method for analyzing OneArmStudy objects.
}
