% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_logic.R
\name{analyze_study_SingleArmStudy}
\alias{analyze_study_SingleArmStudy}
\title{Analyze SingleArmStudy}
\usage{
analyze_study_SingleArmStudy(x, ...)
}
\arguments{
\item{x}{A SingleArmStudy object}

\item{...}{Additional arguments}
}
\value{
The modified SingleArmStudy object with results
}
\description{
Analyze SingleArmStudy objects.
}
