% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_logic.R
\name{analyze_study_TwoArmStudy}
\alias{analyze_study_TwoArmStudy}
\title{Analyze TwoArmStudy}
\usage{
analyze_study_TwoArmStudy(x)
}
\arguments{
\item{x}{A TwoArmStudy object}
}
\value{
The modified TwoArmStudy object with results
}
\description{
S7 method for analyzing TwoArmStudy objects.
}
