% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{apply_format}
\alias{apply_format}
\title{Apply a format specification to values}
\usage{
apply_format(spec, x, align = TRUE)
}
\arguments{
\item{spec}{A FormatSpec object or pattern string}

\item{x}{Numeric vector to format}

\item{align}{Logical, right-align values for tabular display}
}
\value{
Character vector of formatted values
}
\description{
Format numeric values according to the specification.
}
\examples{
\dontrun{
apply_format("xx.x", 12.345) # "12.3"
apply_format("a.xx", c(1, 123)) # "  1.00", "123.00"
}
}
