% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_core.R
\name{apply_subgroups}
\alias{apply_subgroups}
\title{Apply Subgroup Analysis}
\usage{
apply_subgroups(data, subgroup_var, analysis_fn, ...)
}
\arguments{
\item{data}{ADaMData object or data frame}

\item{subgroup_var}{Character string specifying the subgroup variable}

\item{analysis_fn}{Function to perform analysis}

\item{...}{Additional arguments}
}
\value{
A list of AnalysisResults per subgroup
}
\description{
Apply Subgroup Analysis
}
