% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{as_flextable}
\alias{as_flextable}
\title{Convert analysis results to flextable}
\usage{
as_flextable(x, ...)
}
\arguments{
\item{x}{An AnalysisResults or ClinicalTable object}

\item{...}{Additional arguments passed to methods}
}
\value{
A flextable object
}
\description{
Generic function for converting analysis results or clinical tables
to flextable objects for consistent reporting.
}
