% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_engine.R
\name{as_flextable_AnalysisResults}
\alias{as_flextable_AnalysisResults}
\title{Convert AnalysisResults to flextable (S7 Method)}
\usage{
as_flextable_AnalysisResults(x, ...)
}
\arguments{
\item{x}{AnalysisResults object}

\item{...}{Additional arguments}
}
\description{
Convert AnalysisResults to flextable (S7 Method)
}
