% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{as_gt}
\alias{as_gt}
\title{Convert analysis results to gt}
\usage{
as_gt(x, ...)
}
\arguments{
\item{x}{An AnalysisResults object}

\item{...}{Additional arguments passed to methods}
}
\value{
A gt object
}
\description{
Generic function for converting analysis results to gt objects.
}
