% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_engine.R
\name{as_gt_AnalysisResults}
\alias{as_gt_AnalysisResults}
\title{Convert AnalysisResults to gt (S7 Method)}
\usage{
as_gt_AnalysisResults(x, ...)
}
\arguments{
\item{x}{AnalysisResults object}

\item{...}{Additional arguments}
}
\description{
Convert AnalysisResults to gt (S7 Method)
}
