% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_plots.R
\name{bias_plots}
\alias{bias_plots}
\title{Risk of Bias Visualization Functions}
\description{
Functions for visualizing risk of bias assessments using traffic light
plots and summary plots.
}
\references{
McGuinness LA, Higgins JPT (2020). Risk-of-bias VISualization (robvis).
R package version 0.5.1. https://github.com/mcguinlu/robvis
}
