% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_robins_i.R
\name{bias_robins_i}
\alias{bias_robins_i}
\title{ROBINS-I (Risk Of Bias In Non-randomized Studies - of Interventions)}
\description{
Functions for assessing risk of bias in non-randomized studies
using the ROBINS-I tool.
}
\references{
Sterne JAC, et al. (2016). ROBINS-I: a tool for assessing risk of bias
in non-randomized studies of interventions. BMJ, 354, i4919.

IQWiG Allgemeine Methoden Version 10.1, Section 10.1.4.

Cochrane Handbook for Systematic Reviews of Interventions, Chapter 25.
}
