% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{build_layer}
\alias{build_layer}
\title{Build a single layer}
\usage{
build_layer(layer, ...)
}
\arguments{
\item{layer}{An AnalysisLayer object}

\item{data}{Source data frame}

\item{trt_var}{Treatment variable name}
}
\value{
A data frame with layer results
}
\description{
Build a single layer
}
