% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_core.R
\name{calculate_baseline}
\alias{calculate_baseline}
\title{Calculate Baseline Characteristics}
\usage{
calculate_baseline(data, vars)
}
\arguments{
\item{data}{ADaMData object}

\item{vars}{Character vector of variables to analyze}
}
\value{
AnalysisResults object
}
\description{
Calculate Baseline Characteristics
}
