% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{calculate_effective_sample_size}
\alias{calculate_effective_sample_size}
\title{Calculate Effective Sample Size}
\usage{
calculate_effective_sample_size(fit)
}
\arguments{
\item{fit}{A brmsfit object from brms package}
}
\value{
Named numeric vector of ESS values for each parameter
}
\description{
Calculates the effective sample size (ESS) for MCMC parameters.
}
\details{
ESS represents the number of independent samples that would give the same
precision as the correlated MCMC samples. Higher values indicate better
mixing of the chains.
}
\examples{
\dontrun{
ess_values <- calculate_effective_sample_size(fit)
print(ess_values)
}
}
