% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_subgroup.R
\name{calculate_interaction_pvalue_table}
\alias{calculate_interaction_pvalue_table}
\title{Calculate Interaction P-value for Table}
\usage{
calculate_interaction_pvalue_table(
  df,
  subgroup_var,
  endpoint_type,
  time_var,
  event_var_use,
  trt_var
)
}
\arguments{
\item{df}{Data frame}

\item{subgroup_var}{Subgroup variable}

\item{endpoint_type}{"tte" or "binary"}

\item{time_var}{Time variable for TTE}

\item{event_var_use}{Event variable for TTE}

\item{trt_var}{Treatment variable}
}
\value{
Interaction p-value
}
\description{
Calculate Interaction P-value for Table
}
\keyword{internal}
