% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_engine.R
\name{clinical_table_from_results}
\alias{clinical_table_from_results}
\title{Helper to create a ClinicalTable from AnalysisResults}
\usage{
clinical_table_from_results(res, title = "")
}
\arguments{
\item{res}{AnalysisResults object}

\item{title}{Character string for table title}
}
\description{
Helper to create a ClinicalTable from AnalysisResults
}
