% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing_risks.R
\name{CompetingRiskResult}
\alias{CompetingRiskResult}
\title{CompetingRiskResult Class}
\usage{
CompetingRiskResult(
  model = NULL,
  cif_main = data.frame(),
  cif_competing = data.frame(),
  cif_by_treatment = list(),
  treatment_comparison = data.frame(),
  subhazard_ratio = data.frame(),
  main_event = integer(0),
  competing_events = NULL,
  time_points = integer(0),
  n_obs = integer(0),
  n_events = integer(0),
  metadata = list()
)
}
\arguments{
\item{model}{The cmprsk model object}

\item{cif_main}{Data frame with cumulative incidence function for main event}

\item{cif_competing}{Data frame with cumulative incidence function for competing events}

\item{cif_by_treatment}{List of CIFs by treatment group}

\item{treatment_comparison}{Data frame comparing treatments}

\item{subhazard_ratio}{Data frame with subhazard ratios}

\item{main_event}{Integer code for main event of interest}

\item{competing_events}{Integer vector of competing event codes}

\item{time_points}{Numeric vector of time points used}

\item{n_obs}{Number of observations}

\item{n_events}{Vector of event counts by type}

\item{metadata}{List of additional metadata}
}
\value{
A CompetingRiskResult object
}
\description{
An S7 class for storing competing risk analysis results.
}
