% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_classes.R
\name{Configuration}
\alias{Configuration}
\title{Abstract base class for all configuration objects}
\usage{
Configuration(source = "yaml", priority = 0L)
}
\description{
Abstract base class for all configuration objects
}
\keyword{internal}
