% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_tables.R
\name{create_ae_km_plot_for_soc}
\alias{create_ae_km_plot_for_soc}
\title{Create AE KM Plot for a specific SOC}
\usage{
create_ae_km_plot_for_soc(adsl, adae, soc, trt_var = "TRT01A")
}
\arguments{
\item{adsl}{ADSL data frame}

\item{adae}{ADAE data frame}

\item{soc}{SOC value to filter by}

\item{trt_var}{Treatment variable name}
}
\value{
ClinicalPlot object
}
\description{
Create AE KM Plot for a specific SOC
}
