% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_tables.R
\name{create_ae_pt_tables_by_soc}
\alias{create_ae_pt_tables_by_soc}
\title{Create AE PT Tables for all SOCs}
\usage{
create_ae_pt_tables_by_soc(adae, trt_n, trt_var = "TRT01A", autofit = TRUE)
}
\arguments{
\item{adae}{ADAE data frame}

\item{trt_n}{Treatment group counts}

\item{trt_var}{Treatment variable name}

\item{autofit}{Logical, whether to autofit column widths (default: TRUE)}
}
\value{
List of ClinicalTable objects
}
\description{
Create AE PT Tables for all SOCs
}
