% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_tables.R
\name{create_ae_soc_table}
\alias{create_ae_soc_table}
\title{Create AE by System Organ Class Table}
\usage{
create_ae_soc_table(
  adae,
  trt_n,
  title = "Adverse Events by System Organ Class",
  trt_var = "TRT01A",
  autofit = TRUE
)
}
\arguments{
\item{adae}{ADAE data frame}

\item{trt_n}{Treatment group counts}

\item{title}{Table title}

\item{trt_var}{Treatment variable name}
}
\value{
ClinicalTable object
}
\description{
Create AE by System Organ Class Table
}
