% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chef_integration.R
\name{create_chef_endpoint}
\alias{create_chef_endpoint}
\title{Create Chef Endpoint Specification}
\usage{
create_chef_endpoint(
  name,
  variable,
  type = "binary",
  population = NULL,
  strata = character(),
  stats = list(),
  criteria = list()
)
}
\arguments{
\item{name}{Endpoint name}

\item{variable}{Variable name in ADaM dataset}

\item{type}{Endpoint type (e.g., "binary", "continuous", "tte")}

\item{population}{Population filter expression}

\item{strata}{Character vector of stratification variables}

\item{stats}{List of statistical functions to apply}

\item{criteria}{List of inclusion criteria}
}
\value{
A list suitable for chef endpoint definition
}
\description{
Helper to create endpoint specifications compatible with chef pipeline.
}
\examples{
\dontrun{
ep_spec <- create_chef_endpoint(
  name = "Response Rate",
  variable = "AVALC",
  type = "binary",
  strata = c("SEX", "AGEGR1"),
  stats = list(chefStats::n_subj, chefStats::prop_est)
)
}
}
