% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_tables.R
\name{create_common_ae_table}
\alias{create_common_ae_table}
\title{Create Most Common AE Table}
\usage{
create_common_ae_table(
  adae,
  trt_n,
  title = "Most Common Adverse Events",
  n_top = 15,
  trt_var = "TRT01A",
  autofit = TRUE
)
}
\arguments{
\item{adae}{ADAE data frame}

\item{trt_n}{Treatment group counts}

\item{title}{Table title}

\item{n_top}{Number of top PTs to show}

\item{trt_var}{Treatment variable name}
}
\value{
ClinicalTable object
}
\description{
Create Most Common AE Table
}
