% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_conmeds_table}
\alias{create_conmeds_table}
\title{Create Concomitant Medications Table}
\usage{
create_conmeds_table(
  adsl,
  adcm,
  title = "Prior and Concomitant Medications by Class",
  trt_var = "TRT01P",
  class_var = "CMCLAS",
  autofit = TRUE
)
}
\arguments{
\item{adsl}{ADSL data frame}

\item{adcm}{ADCM data frame}

\item{title}{Table title}

\item{trt_var}{Treatment variable name}

\item{class_var}{ATC class/category variable name}
}
\value{
ClinicalTable object
}
\description{
Create Concomitant Medications Table
}
