% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_medical_history_table}
\alias{create_medical_history_table}
\title{Create Medical History Table}
\usage{
create_medical_history_table(
  adsl,
  admh,
  title = "Medical History by Body System",
  trt_var = "TRT01P",
  soc_var = "MHBODSYS",
  autofit = TRUE
)
}
\arguments{
\item{adsl}{ADSL data frame}

\item{admh}{ADMH data frame}

\item{title}{Table title}

\item{trt_var}{Treatment variable name}

\item{soc_var}{SOC variable name for MH}
}
\value{
ClinicalTable object
}
\description{
Create Medical History Table
}
