% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_specification.R
\name{create_prior_specification_set}
\alias{create_prior_specification_set}
\title{Create Prior Specification Set}
\usage{
create_prior_specification_set(priors)
}
\arguments{
\item{priors}{Named list of PriorSpecification objects}
}
\value{
Named list of PriorSpecification objects
}
\description{
Creates a named list of prior specifications for use in Bayesian models.
}
\examples{
\dontrun{
priors <- create_prior_specification_set(list(
  overall = get_default_prior("overall_effect"),
  heterogeneity = get_default_prior("heterogeneity")
))
}
}
