% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_tables.R
\name{create_sae_table}
\alias{create_sae_table}
\title{Create SAE Table}
\usage{
create_sae_table(
  adae,
  trt_n,
  title = "Serious Adverse Events",
  trt_var = "TRT01A",
  autofit = TRUE
)
}
\arguments{
\item{adae}{ADAE data frame}

\item{trt_n}{Treatment group counts}

\item{title}{Table title}

\item{trt_var}{Treatment variable name}
}
\value{
ClinicalTable object
}
\description{
Create SAE Table
}
