% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_logic.R
\name{create_study_report}
\alias{create_study_report}
\title{Create Report from Study}
\usage{
create_study_report(x, title = NULL)
}
\arguments{
\item{x}{Study object}

\item{title}{Character string for report title}
}
\value{
ClinicalReport object
}
\description{
Create Report from Study
}
