% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_bias_adjusted.R
\name{.bias_adjusted_exclude_high}
\alias{.bias_adjusted_exclude_high}
\title{Internal: Exclude high-risk studies method}
\usage{
.bias_adjusted_exclude_high(
  yi,
  sei,
  study_labels,
  include_study,
  meta_result,
  conf_level,
  rob_judgments
)
}
\description{
Internal: Exclude high-risk studies method
}
\keyword{internal}
