% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_bias_adjusted.R
\name{.bias_adjusted_weight_downgrade}
\alias{.bias_adjusted_weight_downgrade}
\title{Internal: Weight downgrade method}
\usage{
.bias_adjusted_weight_downgrade(
  yi,
  sei,
  study_labels,
  study_weights,
  include_study,
  meta_result,
  conf_level,
  rob_judgments
)
}
\description{
Internal: Weight downgrade method
}
\keyword{internal}
