% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.PH_DEFAULTS}
\alias{.PH_DEFAULTS}
\title{Package-wide default values}
\format{
An object of class \code{list} of length 10.
}
\usage{
.PH_DEFAULTS
}
\description{
Central registry of default parameter values used across the package.
These can be overridden using options("pharmhand.\if{html}{\out{<name>}}", value).
}
\keyword{internal}
