% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_plots.R
\docType{data}
\name{.ROB2_DOMAIN_LABELS}
\alias{.ROB2_DOMAIN_LABELS}
\title{RoB 2 Domain Labels}
\format{
An object of class \code{character} of length 5.
}
\usage{
.ROB2_DOMAIN_LABELS
}
\description{
RoB 2 Domain Labels
}
\keyword{internal}
