% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_plots.R
\docType{data}
\name{.ROBINSI_DOMAIN_LABELS}
\alias{.ROBINSI_DOMAIN_LABELS}
\title{ROBINS-I Domain Labels}
\format{
An object of class \code{character} of length 7.
}
\usage{
.ROBINSI_DOMAIN_LABELS
}
\description{
ROBINS-I Domain Labels
}
\keyword{internal}
