% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence_grading.R
\name{evidence_grading}
\alias{evidence_grading}
\title{IQWiG Evidence Grading Implementation}
\description{
Functions for grading evidence according to IQWiG methods
(IQWiG Allgemeine Methoden Version 3.1.4, p.51-56).
}
\references{
IQWiG Allgemeine Methoden Version 3.1.4, Section 3.1, p.51-56.
\url{https://www.iqwig.de/methoden/allgemeine-methoden-version-3-1.pdf}
}
