% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chef_integration.R
\name{extract_chef_metadata}
\alias{extract_chef_metadata}
\title{Extract Metadata from Chef Output}
\usage{
extract_chef_metadata(dt)
}
\arguments{
\item{dt}{Chef output data.table}
}
\value{
List of metadata
}
\description{
Extract Metadata from Chef Output
}
\keyword{internal}
