% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_number}
\alias{format_number}
\title{Format Number with Locale}
\usage{
format_number(x, digits = 2L, locale = get_locale())
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Integer number of decimal places}

\item{locale}{Character locale: "en" (period) or "de" (comma)}
}
\value{
Character formatted number
}
\description{
Helper function to format a number with locale-aware decimal separator.
}
\examples{
format_number(1234.567, 2, "en")
# [1] "1234.57"

format_number(1234.567, 2, "de")
# [1] "1234,57"
}
