% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_presets}
\alias{format_presets}
\alias{fmt_n_pct}
\alias{fmt_mean_sd}
\alias{fmt_median_range}
\alias{fmt_ci}
\alias{fmt_pvalue}
\title{Common Clinical Format Presets}
\usage{
fmt_n_pct()

fmt_mean_sd()

fmt_median_range()

fmt_ci()

fmt_pvalue(threshold = 0.001)
}
\description{
Pre-defined format specifications for common clinical table patterns.
}
\section{Functions}{
\itemize{
\item \code{fmt_n_pct()}: Count with percentage: "n (xx.x\%)"

\item \code{fmt_mean_sd()}: Mean with SD: "xx.x (xx.xx)"

\item \code{fmt_median_range()}: Median with range: "xx.x (xx.x, xx.x)"

\item \code{fmt_ci()}: Confidence interval: "xx.xx (xx.xx, xx.xx)"

\item \code{fmt_pvalue()}: P-value with significance threshold

}}
