% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_spec}
\alias{format_spec}
\title{Create Format Specification}
\usage{
format_spec(
  pattern = "a.a",
  null_format = "--",
  empty_format = "",
  neg_format = "sign"
)
}
\description{
Define numeric formats using pattern strings.
fmt <- format_spec("xx.x")
apply_format(fmt, 12.345) # "12.3"
}
\section{Auto-width with 2 decimals}{

}

