% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{FormatSpec}
\alias{FormatSpec}
\title{FormatSpec S7 Class}
\usage{
FormatSpec(
  pattern = character(0),
  null_format = "--",
  empty_format = "",
  neg_format = "sign",
  int_width = NULL,
  dec_width = NULL,
  auto_int = FALSE,
  auto_dec = FALSE,
  dec_adjust = 0L
)
}
\description{
Internal class for storing format specifications.
}
\keyword{internal}
