% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatting}
\alias{formatting}
\title{Format String Grammar for Clinical Tables}
\description{
A declarative formatting system inspired by Tplyr's f_str() function.
Allows specification of numeric formats using intuitive pattern strings.
}
