% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{generate_word}
\alias{generate_word}
\title{Generate a Word document from a ClinicalReport}
\usage{
generate_word(obj, path, include_title = TRUE, include_toc = TRUE)
}
\arguments{
\item{obj}{A ClinicalReport object}

\item{path}{Character string for output file path}

\item{include_title}{Logical, whether to include study title}

\item{include_toc}{Logical, whether to include table of contents}
}
\value{
The ClinicalReport object (invisibly)
}
\description{
Generate a Word document from a ClinicalReport
}
