% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chef_integration.R
\name{get_chef_stat}
\alias{get_chef_stat}
\title{Register Chef Statistical Functions}
\usage{
get_chef_stat(stat_name)
}
\arguments{
\item{stat_name}{Name of the chefStats function}
}
\value{
A function wrapped with metadata for traceability
}
\description{
Wrapper to access chefStats functions with pharmhand metadata.
}
