% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{get_complete_data}
\alias{get_complete_data}
\title{Get Completed Datasets from Imputation}
\usage{
get_complete_data(imputation_result, action = c("list", "long", "stacked"))
}
\arguments{
\item{imputation_result}{An ImputationResult object}

\item{action}{Character. How to return completed data:
\itemize{
\item "list": Return a list of m data frames (default)
\item "long": Return a single stacked data frame with .imp column
\item "stacked": Alias for "long"
}}
}
\value{
Depending on action: a list of data frames or a single data frame
}
\description{
Extracts completed (imputed) datasets from an ImputationResult object.
}
\examples{
\dontrun{
imp <- perform_multiple_imputation(data, m = 5)

# Get as list
datasets <- get_complete_data(imp, action = "list")
length(datasets) # 5

# Get as stacked data frame
stacked <- get_complete_data(imp, action = "long")
table(stacked$.imp) # Shows imputation numbers
}
}
