% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_api.R
\name{get_performance_setting}
\alias{get_performance_setting}
\title{Get performance setting}
\usage{
get_performance_setting(registry, name, default = NULL)
}
\arguments{
\item{registry}{A ConfigurationRegistry object}

\item{name}{Character string name of the setting}

\item{default}{Default value if setting not found}
}
\value{
The setting value
}
\description{
Retrieves a performance setting value.
}
\examples{
\dontrun{
registry <- load_config()
batch_size <- get_performance_setting(registry, "docx.batch_size", 50)
}
}
