% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_api.R
\name{get_subgroup_config}
\alias{get_subgroup_config}
\title{Get subgroup configuration}
\usage{
get_subgroup_config(registry, name)
}
\arguments{
\item{registry}{A ConfigurationRegistry object}

\item{name}{Character string name of the subgroup}
}
\value{
A SubgroupConfig object or NULL if not found
}
\description{
Retrieves a subgroup configuration by name.
}
\examples{
\dontrun{
registry <- load_config()
config <- get_subgroup_config(registry, "age_groups")
}
}
