% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_core.R
\name{get_subject_var}
\alias{get_subject_var}
\title{Get Subject Variable Name}
\usage{
get_subject_var(data, default = "USUBJID")
}
\arguments{
\item{data}{ADaMData object or data frame}

\item{default}{Default subject variable name for data frames}
}
\value{
Character string with subject variable name
}
\description{
Extract subject ID variable name from ADaMData or return default.
}
\keyword{internal}
