% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_core.R
\name{get_trt_var}
\alias{get_trt_var}
\title{Get Treatment Variable Name}
\usage{
get_trt_var(data, default = "TRT01P")
}
\arguments{
\item{data}{ADaMData object or data frame}

\item{default}{Default treatment variable name for data frames}
}
\value{
Character string with treatment variable name
}
\description{
Extract treatment variable name from ADaMData or return default.
}
\keyword{internal}
