% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{.PH_DEFAULTS}
\alias{\%||\%}
\title{Null Coalescing Operator}
\format{
An object of class \code{list} of length 10.
}
\usage{
.PH_DEFAULTS
}
\description{
Re-export of rlang's null coalescing operator.
See \code{\link[rlang]{\%||\%}} for details.
}
\keyword{internal}
