% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Multiple Imputation Functions}
\description{
Functions for multiple imputation using mice package with
Rubin's rules for pooling results.
}
\references{
IQWiG General Methods 10.3.11 p.233-235
}
