% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{ImputationResult}
\alias{ImputationResult}
\title{ImputationResult Class}
\usage{
ImputationResult(
  mice_object = NULL,
  m = 5L,
  method = "pmm",
  imputed_vars = character(0),
  n_missing = list(),
  original_data = data.frame(),
  metadata = list()
)
}
\arguments{
\item{mice_object}{The mice mids object containing imputed datasets}

\item{m}{Integer number of imputations performed}

\item{method}{Character string or vector specifying imputation method(s)}

\item{imputed_vars}{Character vector of variables that were imputed}

\item{n_missing}{List with count of missing values per variable}

\item{original_data}{The original data frame before imputation}

\item{metadata}{List of additional metadata}
}
\value{
An ImputationResult object
}
\description{
An S7 class for storing multiple imputation results from mice.
}
\examples{
\dontrun{
data <- data.frame(x = c(1, NA, 3), y = c(NA, 2, 3))
result <- perform_multiple_imputation(data, m = 5)
result@m # Number of imputations
result@imputed_vars # Variables that were imputed
}
}
