% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chef_integration.R
\name{list_chef_stats}
\alias{list_chef_stats}
\title{List Available Chef Statistical Functions}
\usage{
list_chef_stats()
}
\value{
Character vector of available chefStats function names
}
\description{
List Available Chef Statistical Functions
}
