% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_api.R
\name{list_populations}
\alias{list_populations}
\title{List available populations}
\usage{
list_populations(registry)
}
\arguments{
\item{registry}{A ConfigurationRegistry object}
}
\value{
Character vector of population names
}
\description{
Returns names of all configured populations.
}
\examples{
\dontrun{
registry <- load_config()
list_populations(registry)
}
}
