% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_api.R
\name{list_subgroups}
\alias{list_subgroups}
\title{List available subgroups}
\usage{
list_subgroups(registry)
}
\arguments{
\item{registry}{A ConfigurationRegistry object}
}
\value{
Character vector of subgroup names
}
\description{
Returns names of all configured subgroups.
}
\examples{
\dontrun{
registry <- load_config()
list_subgroups(registry)
}
}
