% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_bayesian_few.R
\name{meta_bayesian_few}
\alias{meta_bayesian_few}
\title{Bayesian Meta-Analysis for Few Studies}
\description{
Specialized Bayesian meta-analysis functions for situations
with few studies, using appropriate priors and sensitivity analysis.
}
\references{
IQWiG General Methods 10.3.7 p.224-227
}
