% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm.R
\name{mmrm}
\alias{mmrm}
\title{Mixed Model Repeated Measures (MMRM)}
\description{
Functions for mixed model repeated measures analysis of
longitudinal patient-reported outcomes (PRO) data.
}
\references{
IQWiG General Methods 10.3.9 p.229-232
}
