% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm.R
\name{MMRMResult}
\alias{MMRMResult}
\title{MMRMResult Class}
\usage{
MMRMResult(
  model = NULL,
  coefficients = integer(0),
  se = integer(0),
  ci = NULL,
  p_values = integer(0),
  df = integer(0),
  sigma = integer(0),
  log_likelihood = integer(0),
  aic = integer(0),
  bic = integer(0),
  covariance = "unstructured",
  df_adjustment = "Kenward-Roger",
  n_obs = integer(0),
  n_subjects = integer(0),
  metadata = list()
)
}
\arguments{
\item{model}{The mmrm model object}

\item{coefficients}{Named numeric vector of model coefficients}

\item{se}{Standard errors for coefficients}

\item{ci}{Matrix of confidence intervals (2 columns: lower, upper)}

\item{p_values}{P-values for coefficients}

\item{df}{Degrees of freedom for coefficients}

\item{sigma}{Residual standard error}

\item{log_likelihood}{Log-likelihood value}

\item{aic}{Akaike Information Criterion}

\item{bic}{Bayesian Information Criterion}

\item{covariance}{Covariance structure used}

\item{df_adjustment}{Degrees of freedom adjustment method}

\item{n_obs}{Number of observations}

\item{n_subjects}{Number of subjects}

\item{metadata}{List of additional metadata}
}
\value{
An MMRMResult object
}
\description{
An S7 class for storing MMRM analysis results.
}
