% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_api.R
\name{parse_config_registry}
\alias{parse_config_registry}
\title{Parse configuration from YAML list}
\usage{
parse_config_registry(config)
}
\arguments{
\item{config}{List from yaml::read_yaml()}
}
\value{
A ConfigurationRegistry object
}
\description{
Internal function to parse YAML configuration list into
ConfigurationRegistry object.
}
\keyword{internal}
