% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{ph_abort}
\alias{ph_abort}
\title{Abort with Error}
\usage{
ph_abort(...)
}
\arguments{
\item{...}{Arguments passed to rlang::abort()}
}
\description{
Internal wrapper for rlang::abort() that provides consistent error formatting.
}
\keyword{internal}
