% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{ph_inform}
\alias{ph_inform}
\title{Inform User}
\usage{
ph_inform(...)
}
\arguments{
\item{...}{Arguments passed to message()}
}
\description{
Internal wrapper for message() that provides consistent message formatting.
}
\keyword{internal}
