% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{ph_warn}
\alias{ph_warn}
\title{Issue Warning}
\usage{
ph_warn(...)
}
\arguments{
\item{...}{Arguments passed to warning()}
}
\description{
Internal wrapper for warning() that provides consistent warning formatting.
}
\keyword{internal}
