% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing_risks.R
\name{plot_cif}
\alias{plot_cif}
\title{Plot Cumulative Incidence Function}
\usage{
plot_cif(result, title = "Cumulative Incidence Function")
}
\arguments{
\item{result}{A CompetingRiskResult object}

\item{title}{Plot title}
}
\value{
A ClinicalPlot object
}
\description{
Creates a plot of the cumulative incidence function from competing risk analysis.
}
\examples{
\dontrun{
plot <- plot_cif(result, title = "Cumulative Incidence Function")
}
}
