% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_diagnostics.R
\name{plot_imputation_convergence}
\alias{plot_imputation_convergence}
\title{Plot Imputation Convergence}
\usage{
plot_imputation_convergence(
  imputation_result,
  vars = NULL,
  title = "Imputation Convergence"
)
}
\arguments{
\item{imputation_result}{An ImputationResult object}

\item{vars}{Character vector of variables to plot. If NULL, plots all
imputed variables.}

\item{title}{Plot title}
}
\value{
A ClinicalPlot object
}
\description{
Creates trace plots showing MCMC convergence of the imputation algorithm.
}
\examples{
\dontrun{
imp <- perform_multiple_imputation(data, m = 5)
plot_imputation_convergence(imp)
}
}
