% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_diagnostics.R
\name{plot_imputation_distributions}
\alias{plot_imputation_distributions}
\title{Plot Imputation Distributions}
\usage{
plot_imputation_distributions(
  imputation_result,
  vars = NULL,
  title = "Observed vs Imputed Distributions"
)
}
\arguments{
\item{imputation_result}{An ImputationResult object}

\item{vars}{Character vector of variables to plot. If NULL, plots all
imputed numeric variables.}

\item{title}{Plot title}
}
\value{
A ClinicalPlot object
}
\description{
Compares distributions of observed vs imputed values.
}
\examples{
\dontrun{
imp <- perform_multiple_imputation(data, m = 5)
plot_imputation_distributions(imp)
}
}
