% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_diagnostics.R
\name{plot_missing_pattern}
\alias{plot_missing_pattern}
\title{Plot Missing Data Pattern}
\usage{
plot_missing_pattern(x, title = "Missing Data Pattern")
}
\arguments{
\item{x}{A data frame or ImputationResult object}

\item{title}{Plot title}
}
\value{
A ClinicalPlot object
}
\description{
Visualizes the pattern of missing data in a dataset.
}
\examples{
data <- data.frame(a = c(1, NA, 3), b = c(NA, 2, 3))
plot_missing_pattern(data)
}
