% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{plot_rmst}
\alias{plot_rmst}
\title{Plot RMST Results}
\usage{
plot_rmst(result, title = "Restricted Mean Survival Time by Treatment Group")
}
\arguments{
\item{result}{An RMSTResult object}

\item{title}{Plot title}
}
\value{
A ClinicalPlot object
}
\description{
Creates a plot of RMST results by treatment group.
}
\examples{
\dontrun{
plot <- plot_rmst(result, title = "RMST by Treatment Group")
}
}
