% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_bias_adjusted.R
\name{print.BiasAdjustedSummary}
\alias{print.BiasAdjustedSummary}
\title{Print Method for BiasAdjustedSummary}
\usage{
\method{print}{BiasAdjustedSummary}(x, ...)
}
\arguments{
\item{x}{A BiasAdjustedSummary object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print Method for BiasAdjustedSummary
}
\keyword{internal}
