% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence_grading.R
\name{print.EvidenceGrade}
\alias{print.EvidenceGrade}
\title{Print Method for EvidenceGrade}
\usage{
\method{print}{EvidenceGrade}(x, ...)
}
\arguments{
\item{x}{An EvidenceGrade object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print Method for EvidenceGrade
}
\keyword{internal}
