% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_rob2.R
\name{print.RoB2Result}
\alias{print.RoB2Result}
\title{Print Method for RoB2Result}
\usage{
\method{print}{RoB2Result}(x, ...)
}
\arguments{
\item{x}{A RoB2Result object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print Method for RoB2Result
}
\keyword{internal}
