% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_robins_i.R
\name{print.ROBINSIResult}
\alias{print.ROBINSIResult}
\title{Print Method for ROBINSIResult}
\usage{
\method{print}{ROBINSIResult}(x, ...)
}
\arguments{
\item{x}{A ROBINSIResult object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print Method for ROBINSIResult
}
\keyword{internal}
