% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst}
\alias{rmst}
\title{Restricted Mean Survival Time (RMST)}
\description{
Functions for restricted mean survival time analysis using
Kaplan-Meier estimation with treatment group comparisons.
}
\references{
IQWiG General Methods 10.3.8 p.227-229
}
