% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{RMSTResult}
\alias{RMSTResult}
\title{RMSTResult Class}
\usage{
RMSTResult(
  rmst_by_group = data.frame(),
  rmst_difference = integer(0),
  se_difference = integer(0),
  ci = NULL,
  p_value = integer(0),
  tau = integer(0),
  treatment_comparison = data.frame(),
  n_obs = integer(0),
  n_events = integer(0),
  metadata = list()
)
}
\arguments{
\item{rmst_by_group}{Data frame with RMST estimates by treatment group}

\item{rmst_difference}{Numeric. Difference in RMST between groups}

\item{se_difference}{Standard error of the difference}

\item{ci}{Numeric vector c(lower, upper) confidence interval for difference}

\item{p_value}{P-value for treatment difference test}

\item{tau}{Numeric. Time restriction used for RMST calculation}

\item{treatment_comparison}{Data frame with treatment comparison results}

\item{n_obs}{Number of observations}

\item{n_events}{Vector of event counts by group}

\item{metadata}{List of additional metadata}
}
\value{
An RMSTResult object
}
\description{
An S7 class for storing RMST analysis results.
}
