% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{StudySet}
\alias{StudySet}
\title{StudySet Class}
\usage{
StudySet(
  studies = list(),
  endpoint = NULL,
  comparison_type = "direct",
  common_comparator = NULL,
  characteristics = structure(list(), names = character(0), row.names = integer(0), class
    = "data.frame"),
  metadata = list()
)
}
\arguments{
\item{studies}{List of Study objects}

\item{endpoint}{Endpoint object being synthesized}

\item{comparison_type}{Type of comparison: "direct", "indirect", "network"}

\item{common_comparator}{Common comparator for indirect/network comparisons}

\item{characteristics}{Data frame of study-level characteristics}

\item{metadata}{List of additional metadata}
}
\value{
A StudySet object
}
\description{
An S7 class for representing a collection of studies for evidence synthesis
(meta-analysis, indirect comparison, network meta-analysis).
}
\examples{
\dontrun{
study_set <- StudySet(
  studies = list(study1, study2, study3),
  endpoint = os_endpoint,
  comparison_type = "direct"
)
}
}
