% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generics.R
\name{summarize_content}
\alias{summarize_content}
\title{Generate a summary of clinical content}
\usage{
summarize_content(x, ...)
}
\arguments{
\item{x}{A clinical content object}

\item{...}{Additional arguments passed to methods}
}
\value{
A list with summary information
}
\description{
Generic function for generating summary information about
clinical content objects.
}
\examples{
\dontrun{
summary_info <- summarize_content(clinical_table)
}
}
