% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_specification.R
\name{summarize_prior_specification}
\alias{summarize_prior_specification}
\title{Summarize Prior Specification}
\usage{
summarize_prior_specification(prior)
}
\arguments{
\item{prior}{A PriorSpecification object}
}
\value{
A data frame with prior summary information
}
\description{
Creates a summary data frame from a prior specification.
}
\examples{
\dontrun{
prior <- create_prior_specification("normal", list(mean = 0, sd = 1))
summary <- summarize_prior_specification(prior)
print(summary)
}
}
