% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_engine.R
\name{write_docx_ClinicalReport}
\alias{write_docx_ClinicalReport}
\title{Write ClinicalReport to Word (S7 Method)}
\usage{
write_docx_ClinicalReport(x, path, add_toc = FALSE, ...)
}
\arguments{
\item{x}{ClinicalReport object}

\item{path}{File path}

\item{...}{Additional arguments}
}
\description{
Write ClinicalReport to Word (S7 Method)
}
