% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{AnalysisResults}
\alias{AnalysisResults}
\title{AnalysisResults Class}
\usage{
AnalysisResults(
  stats = data.frame(),
  type = "",
  groupings = list(),
  metadata = list()
)
}
\arguments{
\item{stats}{A data frame containing the statistical results}

\item{type}{Character string for result type}

\item{groupings}{List of grouping variables used}

\item{metadata}{List of additional metadata}
}
\value{
An AnalysisResults object
}
\description{
Container for pre-calculated statistics.
}
