% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_logic.R
\name{analyze_study}
\alias{analyze_study}
\title{Analyze Study (S7 Method)}
\usage{
analyze_study(x, ...)
}
\arguments{
\item{x}{Study object (SingleArmStudy, TwoArmStudy, or MultiArmStudy)}

\item{...}{Additional arguments}
}
\description{
Analyze Study (S7 Method)
}
