% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{apply_composite}
\alias{apply_composite}
\title{Apply a composite format}
\usage{
apply_composite(fmt, ...)
}
\arguments{
\item{fmt}{CompositeFormat object}

\item{...}{Named values matching template placeholders}
}
\value{
Formatted character string
}
\description{
Apply a composite format
}
