% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{calculate_heterogeneity}
\alias{calculate_heterogeneity}
\title{Calculate Heterogeneity Statistics}
\usage{
calculate_heterogeneity(yi, sei, method = c("REML", "DL", "PM", "ML"))
}
\arguments{
\item{yi}{Numeric vector of effect estimates}

\item{sei}{Numeric vector of standard errors}

\item{method}{Character. tau2 estimation method. Default: "REML"}
}
\value{
A list with components:
\describe{
\item{Q}{Cochran's Q statistic}
\item{Q_df}{Degrees of freedom for Q test}
\item{Q_pvalue}{P-value for Q test}
\item{I2}{I-squared heterogeneity percentage (0-100)}
\item{I2_ci}{95 percent CI for I-squared}
\item{H2}{H-squared statistic}
\item{H}{H statistic (sqrt of H2)}
\item{tau2}{Between-study variance estimate}
\item{tau}{Between-study standard deviation}
\item{method}{Estimation method used}
\item{k}{Number of studies}
\item{interpretation}{Verbal interpretation of I2 level}
}
}
\description{
Calculates Q, I2, tau2, and H2 statistics for meta-analysis heterogeneity.
}
\examples{
# Calculate heterogeneity for 5 studies
yi <- c(0.5, 0.8, 0.3, 1.0, 0.6)
sei <- c(0.2, 0.2, 0.2, 0.2, 0.2)
het <- calculate_heterogeneity(yi, sei, method = "REML")
het$Q
het$I2
het$tau2
het$interpretation
}
