% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_engine.R
\name{create_clinical_table}
\alias{create_clinical_table}
\title{Helper to create a ClinicalTable from AnalysisResults}
\usage{
create_clinical_table(res, title = "")
}
\arguments{
\item{res}{AnalysisResults object}

\item{title}{Character string for table title}
}
\description{
Helper to create a ClinicalTable from AnalysisResults
}
