% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_conmeds_table}
\alias{create_conmeds_table}
\title{Create Concomitant Medications Table}
\usage{
create_conmeds_table(
  data,
  adsl,
  title = "Prior and Concomitant Medications by Class",
  trt_var = "TRT01P",
  class_var = "CMCLAS",
  autofit = TRUE
)
}
\arguments{
\item{data}{ADCM data frame or ADaMData object}

\item{adsl}{ADSL data frame or ADaMData object for denominators}

\item{title}{Table title}

\item{trt_var}{Treatment variable name}

\item{class_var}{ATC class/category variable name}

\item{autofit}{Logical, whether to autofit column widths (default: TRUE)}
}
\value{
ClinicalTable object
}
\description{
Create Concomitant Medications Table
}
