% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_km_plot}
\alias{create_km_plot}
\title{Create Kaplan-Meier Plot}
\usage{
create_km_plot(
  data,
  time_var,
  event_var,
  trt_var,
  title = "Kaplan-Meier Plot",
  xlab = "Time",
  ylab = "Survival Probability",
  risk_table = FALSE
)
}
\arguments{
\item{data}{Data frame containing survival data}

\item{time_var}{Time variable name}

\item{event_var}{Event variable name (1=event, 0=censor)}

\item{trt_var}{Treatment variable name}

\item{title}{Plot title}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{risk_table}{Logical, include risk table below. Requires patchwork.}
}
\value{
A ClinicalPlot object
}
\description{
Generates a standard Kaplan-Meier plot using ggplot2 and survival.
}
