% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to check}

\item{y}{Second value to return if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Return first value if not NULL, otherwise return second value.
This is a common pattern in functional programming for providing defaults.
}
\examples{
\dontrun{
NULL \%||\% "default"  # "default"
"value" \%||\% "default"  # "value"
}
}
\keyword{internal}
